@@ -2,55 +2,55 @@
   "name": "luma-co-app",
   "version": "1.0.0",
   "scripts": {
-    "start": "react-scripts start"
+    "start": "react-scripts start",
+    "server": "node backend/server.js",
+    "server-dev": "nodemon backend/server.js",
+    "dev": "concurrently \"npm run start\" \"npm run server-dev\"",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  "devDependencies": {
-    "autoprefixer": "^10.4.21",
-    "concurrently": "^9.1.2",
-    "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.6.3",
-    "postcss": "^8.5.3",
-    "react-scripts": "^5.0.1",
-    "tailwindcss": "^4.0.14",
-    "typescript": "^5.8.2",
-    "webpack-cli": "^6.0.1"
-  },
+  "main": "backend/server.js",
   "dependencies": {
-    "aws-sdk": "^2.1692.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@graphique/geom-line": "^3.0.0",
+    "@heroicons/react": "^2.2.0",
+    "@mui/material": "^6.4.8",
     "axios": "^1.8.4",
-    "bcryptjs": "^3.0.2",
-    "bootstrap": "^5.3.3",
+    "bcrypt": "^5.1.1",
     "chart.js": "^4.4.8",
     "cors": "^2.8.5",
+    "css-tree": "^3.1.0",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
     "framer-motion": "^12.5.0",
-    "i18next": "^24.2.3",
+    "html2canvas": "^1.4.1",
+    "imagemagick": "^0.1.3",
     "jsonwebtoken": "^9.0.2",
     "jspdf": "^3.0.1",
-    "moment": "^2.30.1",
-    "mongoose": "^8.12.1",
-    "mysql2": "^3.14.0",
+    "jspdf-autotable": "^5.0.2",
+    "lucide-react": "^0.483.0",
+    "mongoose": "^8.12.2",
     "nodemailer": "^6.10.0",
-    "pdf-lib": "^1.17.1",
-    "react": "^18.3.1",
-    "react-bootstrap": "^2.10.9",
+    "react": "^19.0.0",
     "react-chartjs-2": "^5.3.0",
-    "react-datepicker": "^8.2.1",
-    "react-dom": "^18.3.1",
-    "react-i18next": "^15.4.1",
+    "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
-    "react-plotly.js": "^2.6.0",
-    "react-router-dom": "^7.4.0",
-    "react-select": "^5.10.1",
-    "react-table": "^7.8.0",
-    "react-toastify": "^11.0.5",
+    "react-router-dom": "^6.30.0",
+    "react-toastify": "^10.0.6",
+    "redux": "^4.2.1",
     "sequelize": "^6.37.6",
-    "socket.io-client": "^4.8.1",
-    "web-push": "^3.6.7",
-    "web-vitals": "^4.2.4",
-    "webpack": "5",
-    "webpack-dev-server": "4"
+    "socket.io-client": "^4.8.1"
+  },
+  "devDependencies": {
+    "@tailwindcss/postcss": "^4.0.15",
+    "autoprefixer": "^10.4.21",
+    "concurrently": "^9.1.2",
+    "nodemon": "^3.1.9",
+    "postcss": "^8.5.3",
+    "postcss-loader": "^8.1.1",
+    "tailwindcss": "^4.0.15"
   },
   "browserslist": {
     "production": [
@@ -63,10 +63,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
-  "main": "index.js",
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
+  }
 }
