@@ -14,15 +14,15 @@ oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")
 router = APIRouter()
 
 # Schéma de l'utilisateur
-class User(BaseModel):
+class User(BaseModel): 
     username: str
     password: str
 
 # Base de données simulée des utilisateurs (avec leurs rôles)
 users_db = {
-    "admin@luma.com": {"password": "admi123@", "role": "admin"},
-    "commercial@example.com": {"password": "commercialpassword", "role": "commercial"},
-    "rh@example.com": {"password": "rhpassword", "role": "rh"},
+    "admin@luma.com": {"password": "admin123", "role": "admin"},
+    "commercial@luma.com": {"password": "commercialpassword", "role": "commercial"},
+    "rh@luma.com": {"password": "rhpassword", "role": "rh"},
 }
 
 # Fonction pour vérifier l'authentification et le rôle
@@ -57,9 +57,15 @@ def get_current_user(token: str = Depends(oauth2_scheme)):
 # Définir les permissions en fonction des rôles
 def check_permissions(role: str, endpoint: str):
     permissions = {
-        "admin": ["dashboard", "clients", "devis", "factures", "chats", "reglages", "contrats", "nominas", "recrutement", "conges", "employes"],
-        "commercial": ["dashboard", "clients", "devis", "factures", "chats", "reglages"],
-        "rh": ["dashboard", "clients", "devis", "factures", "chats", "reglages", "contrats", "nominas", "recrutement", "conges", "employes"]
+        "admin": [
+            "dashboard", "clients", "devis", "factures", "chats", "reglages", "contrats", "nominas", "recrutement", "conges", "employes"
+        ],
+        "commercial": [
+            "dashboard", "clients", "devis", "factures", "chats", "reglages"
+        ],
+        "rh": [
+            "dashboard", "clients", "devis", "factures", "chats", "reglages", "contrats", "nominas", "recrutement", "conges", "employes"
+        ]
     }
     
     if endpoint not in permissions.get(role, []):
@@ -76,7 +82,48 @@ def clients(current_user: dict = Depends(get_current_user)):
     check_permissions(current_user["role"], "clients")
     return {"msg": "Here are the clients!"}
 
+@router.get("/devis")
+def devis(current_user: dict = Depends(get_current_user)):
+    check_permissions(current_user["role"], "devis")
+    return {"msg": "Here are the quotes!"}
+
+@router.get("/factures")
+def factures(current_user: dict = Depends(get_current_user)):
+    check_permissions(current_user["role"], "factures")
+    return {"msg": "Here are the invoices!"}
+
+@router.get("/chats")
+def chats(current_user: dict = Depends(get_current_user)):
+    check_permissions(current_user["role"], "chats")
+    return {"msg": "Here are the chats!"}
+
+@router.get("/reglages")
+def reglages(current_user: dict = Depends(get_current_user)):
+    check_permissions(current_user["role"], "reglages")
+    return {"msg": "Here are the settings!"}
+
 @router.get("/contrats")
 def contrats(current_user: dict = Depends(get_current_user)):
     check_permissions(current_user["role"], "contrats")
     return {"msg": "Here are the contracts!"}
+
+@router.get("/nominas")
+def nominas(current_user: dict = Depends(get_current_user)):
+    check_permissions(current_user["role"], "nominas")
+    return {"msg": "Here are the payrolls!"}
+
+@router.get("/recrutement")
+def recrutement(current_user: dict = Depends(get_current_user)):
+    check_permissions(current_user["role"], "recrutement")
+    return {"msg": "Here is the recruitment section!"}
+
+@router.get("/conges")
+def conges(current_user: dict = Depends(get_current_user)):
+    check_permissions(current_user["role"], "conges")
+    return {"msg": "Here is the leave section!"}
+
+@router.get("/employes")
+def employes(current_user: dict = Depends(get_current_user)):
+    check_permissions(current_user["role"], "employes")
+    return {"msg": "Here is the employees section!"}
+
