@@ -1,6 +1,8 @@
-import ClientModel from "./Client.js";
+// index.js
+import ClientModels from "./Client.js";
 
 export function initModels(sequelize, DataTypes) {
-  const Client = ClientModel(sequelize, DataTypes);
+  // Utilisation du bon nom : ClientModels (avec le 's')
+  const Client = ClientModels(sequelize, DataTypes);
   return { Client };
 }
